Add default plotting description for Collections (#634)
* fix extra traces edge cases for plotly

* fix trace non standard trace merging

* match args for all extra traces

* don't override opacity

* fix mag display only if mag style

* reorder imports

* draft

* pull from

* fixes + add arrow offset

* pylint

* add collection default description

* allow style_description= "sometext"

* rely on obj style directly

* only update mag mesh if mesh3d

* make draw func public

* fix collection description

* refix coll desc

* fix tests

* pylint

* typo

* refactor

* include mag arrows in pre-orientation routine

* refactor into traces_core

* don't colorize magnet orientation

* add some annotations

* fix bypassing dimension and excitation checks

---------
