+++ magpylib/_src/display/traces_generic.py
@@ -7,62 +7,46 @@
 import numbers
 import warnings
 from collections import Counter
-from itertools import combinations
 from itertools import cycle
 from typing import Tuple
 
 import numpy as np
-from scipy.spatial import distance
-from scipy.spatial.transform import Rotation as RotScipy
 
 import magpylib as magpy
 from magpylib._src.defaults.defaults_classes import default_settings as Config
 from magpylib._src.defaults.defaults_utility import ALLOWED_LINESTYLES
 from magpylib._src.defaults.defaults_utility import ALLOWED_SYMBOLS
 from magpylib._src.defaults.defaults_utility import linearize_dict
-from magpylib._src.display.sensor_mesh import get_sensor_mesh
-from magpylib._src.display.traces_base import make_Arrow as make_BaseArrow
-from magpylib._src.display.traces_base import make_Cuboid as make_BaseCuboid
-from magpylib._src.display.traces_base import (
-    make_CylinderSegment as make_BaseCylinderSegment,
-)
-from magpylib._src.display.traces_base import make_Ellipsoid as make_BaseEllipsoid
-from magpylib._src.display.traces_base import make_Prism as make_BasePrism
-from magpylib._src.display.traces_base import make_Pyramid as make_BasePyramid
-from magpylib._src.display.traces_base import make_Tetrahedron as make_BaseTetrahedron
-from magpylib._src.display.traces_base import (
-    make_TriangularMesh as make_BaseTriangularMesh,
-)
-from magpylib._src.display.traces_utility import draw_arrow_from_vertices
-from magpylib._src.display.traces_utility import draw_arrow_on_circle
 from magpylib._src.display.traces_utility import draw_arrowed_line
 from magpylib._src.display.traces_utility import get_flatten_objects_properties
+from magpylib._src.display.traces_utility import get_label
 from magpylib._src.display.traces_utility import get_rot_pos_from_path
 from magpylib._src.display.traces_utility import get_scene_ranges
 from magpylib._src.display.traces_utility import getColorscale
 from magpylib._src.display.traces_utility import getIntensity
 from magpylib._src.display.traces_utility import group_traces
-from magpylib._src.display.traces_utility import merge_mesh3d
-from magpylib._src.display.traces_utility import merge_traces
 from magpylib._src.display.traces_utility import place_and_orient_model3d
 from magpylib._src.display.traces_utility import slice_mesh_from_colorscale
-from magpylib._src.display.traces_utility import triangles_area
 from magpylib._src.style import DefaultMarkers
 from magpylib._src.utility import format_obj_input
-from magpylib._src.utility import unit_prefix
 
 
 class MagpyMarkers:
     """A class that stores markers 3D-coordinates."""
 
     def __init__(self, *markers):
-        self.style = DefaultMarkers()
+        self._style = DefaultMarkers()
         self.markers = np.array(markers)
 
-    def _draw_func(self, style=None, **kwargs):
+    @property
+    def style(self):
+        """Style property"""
+        return self._style
+
+    def get_trace(self, **kwargs):
         """Create the plotly mesh3d parameters for a Sensor object in a dictionary based on the
         provided arguments."""
-        style = self.style if style is None else style
+        style = self.style
         x, y, z = self.markers.T
         marker_kwargs = {
             f"marker_{k}": v
@@ -82,551 +66,8 @@ class MagpyMarkers:
         }
         default_name = "Marker" if len(x) == 1 else "Markers"
         default_suffix = "" if len(x) == 1 else f" ({len(x)} points)"
-        update_trace_name(trace, default_name, default_suffix, style)
+        trace["name"] = get_label(self, default_suffix, default_name)
         return trace
-
-
-def make_DefaultTrace(
-    obj,
-    style=None,
-    **kwargs,
-) -> dict:
-    """
-    Creates the plotly scatter3d parameters for an object with no specifically supported
-    representation. The object will be represented by a scatter point and text above with object
-    name.
-    """
-    style = obj.style if style is None else style
-    trace = {
-        "type": "scatter3d",
-        "x": [0.0],
-        "y": [0.0],
-        "z": [0.0],
-        "mode": "markers+text",
-        "marker_size": 10,
-        "marker_color": style.color,
-        "marker_symbol": "diamond",
-    }
-    update_trace_name(trace, f"{type(obj).__name__}", "", style)
-    trace["text"] = trace["name"]
-    return {**trace, **kwargs}
-
-
-def make_Line(
-    obj,
-    style=None,
-    **kwargs,
-) -> dict:
-    """
-    Creates the plotly scatter3d parameters for a Line current in a dictionary based on the
-    provided arguments.
-    """
-    default_suffix = (
-        f" ({unit_prefix(obj.current)}A)"
-        if obj.current is not None
-        else " (Current not initialized)"
-    )
-    style = obj.style if style is None else style
-    traces = []
-    for kind in ("arrow", "line"):
-        kind_style = getattr(style, kind)
-        if kind_style.show:
-            color = style.color if kind_style.color is None else kind_style.color
-            if kind == "arrow":
-                x, y, z = draw_arrow_from_vertices(
-                    obj.vertices,
-                    obj.current,
-                    kind_style.size,
-                    arrow_pos=style.arrow.offset,
-                    include_line=False,
-                ).T
-            else:
-                x, y, z = obj.vertices.T
-            trace = {
-                "type": "scatter3d",
-                "x": x,
-                "y": y,
-                "z": z,
-                "mode": "lines",
-                "line_width": kind_style.width,
-                "line_dash": kind_style.style,
-                "line_color": color,
-            }
-            update_trace_name(trace, "Line", default_suffix, style)
-            traces.append({**trace, **kwargs})
-    return traces
-
-
-def make_Loop(
-    obj,
-    style=None,
-    vert_num=72,
-    **kwargs,
-):
-    """
-    Creates the plotly scatter3d parameters for a Loop current in a dictionary based on the
-    provided arguments.
-    """
-    default_suffix = (
-        f" ({unit_prefix(obj.current)}A)"
-        if obj.current is not None
-        else " (Current not initialized)"
-    )
-    style = obj.style if style is None else style
-    traces = []
-    for kind in ("arrow", "line"):
-        kind_style = getattr(style, kind)
-        if kind_style.show:
-            color = style.color if kind_style.color is None else kind_style.color
-
-            if kind == "arrow":
-                angle_pos_deg = 360 * np.round(style.arrow.offset * vert_num) / vert_num
-                vertices = draw_arrow_on_circle(
-                    np.sign(obj.current),
-                    obj.diameter,
-                    style.arrow.size,
-                    angle_pos_deg=angle_pos_deg,
-                )
-                x, y, z = vertices.T
-            else:
-                t = np.linspace(0, 2 * np.pi, vert_num)
-                x = np.cos(t) * obj.diameter / 2
-                y = np.sin(t) * obj.diameter / 2
-                z = np.zeros(x.shape)
-            trace = {
-                "type": "scatter3d",
-                "x": x,
-                "y": y,
-                "z": z,
-                "mode": "lines",
-                "line_width": kind_style.width,
-                "line_dash": kind_style.style,
-                "line_color": color,
-            }
-            update_trace_name(trace, "Loop", default_suffix, style)
-            traces.append({**trace, **kwargs})
-    return traces
-
-
-def make_Dipole(
-    obj,
-    style=None,
-    autosize=None,
-    **kwargs,
-) -> dict:
-    """
-    Create the plotly mesh3d parameters for a dipole in a dictionary based on the
-    provided arguments.
-    """
-    style = obj.style if style is None else style
-    moment = obj.moment
-    moment_mag = np.linalg.norm(moment)
-    size = style.size
-    if autosize is not None:
-        size *= autosize
-    trace = make_BaseArrow(
-        "plotly-dict",
-        base=10,
-        diameter=0.3 * size,
-        height=size,
-        pivot=style.pivot,
-        color=style.color,
-    )
-    default_suffix = f" (moment={unit_prefix(moment_mag)}mT mm³)"
-    update_trace_name(trace, "Dipole", default_suffix, style)
-    nvec = np.array(moment) / moment_mag
-    zaxis = np.array([0, 0, 1])
-    cross = np.cross(nvec, zaxis)
-    n = np.linalg.norm(cross)
-    if n == 0:
-        n = 1
-        cross = np.array([-np.sign(nvec[-1]), 0, 0])
-    dot = np.dot(nvec, zaxis)
-    t = np.arccos(dot)
-    vec = -t * cross / n
-    mag_orient = RotScipy.from_rotvec(vec)
-    trace = place_and_orient_model3d(trace, orientation=mag_orient, **kwargs)
-    return {**trace, **kwargs}
-
-
-def get_closest_vertices(faces_subsets, vertices):
-    """Get closest pairs of points between disconnected subsets of faces indices"""
-    nparts = len(faces_subsets)
-    inds_subsets = [np.unique(v) for v in faces_subsets]
-    closest_verts_list = []
-    if nparts > 1:
-        connected = [np.min(inds_subsets[0])]
-        while len(connected) < nparts:
-            prev_min = float("inf")
-            for i in connected:
-                for j in range(nparts):
-                    if j not in connected:
-                        tr1, tr2 = inds_subsets[i], inds_subsets[j]
-                        c1, c2 = vertices[tr1], vertices[tr2]
-                        dist = distance.cdist(c1, c2)
-                        i1, i2 = divmod(dist.argmin(), dist.shape[1])
-                        min_dist = dist[i1, i2]
-                        if min_dist < prev_min:
-                            prev_min = min_dist
-                            closest_verts = [c1[i1], c2[i2]]
-                            connected_ind = j
-            connected.append(connected_ind)
-            closest_verts_list.append(closest_verts)
-    return np.array(closest_verts_list)
-
-
-def make_mesh_lines(
-    obj,
-    pos_orient_inds,
-    mode,
-    style=None,
-    **kwargs,
-):
-    """Draw mesh lines and vertices"""
-    # pylint: disable=protected-access
-    kwargs.pop("color", None)
-    style = obj.style if style is None else style
-    mesh = getattr(style.mesh, mode)
-    marker, line = mesh.marker, mesh.line
-    tr, vert = obj.faces, obj.vertices
-    if mode == "disconnected":
-        subsets = obj.get_faces_subsets()
-        lines = get_closest_vertices(subsets, vert)
-    else:
-        if mode == "selfintersecting":
-            tr = obj.faces[obj.get_selfintersecting_faces()]
-        edges = np.concatenate([tr[:, 0:2], tr[:, 1:3], tr[:, ::2]], axis=0)
-        if mode == "open":
-            edges = obj.get_open_edges()
-        else:
-            edges = np.unique(edges, axis=0)
-        lines = vert[edges]
-
-    out = {}
-    if lines.size != 0:
-        label = f"{obj}" if style.label is None else style.label
-        lines = np.insert(lines, 2, None, axis=1).reshape(-1, 3)
-        traces = []
-        for ind in pos_orient_inds:
-            x, y, z = (obj._orientation[ind].apply(lines) + obj._position[ind]).T
-            trace = {
-                "type": "scatter3d",
-                "x": x,
-                "y": y,
-                "z": z,
-                "marker_color": marker.color,
-                "marker_size": marker.size,
-                "marker_symbol": marker.symbol,
-                "line_color": line.color,
-                "line_width": line.width,
-                "line_dash": line.style,
-                "legendgroup": f"{obj}{mode}edges",
-                "name": f"{label} - {mode}-edges",
-            }
-            traces.append(trace)
-        out = {**merge_traces(*traces)[0], **kwargs}
-    return out
-
-
-def make_triangle_orientations(
-    obj,
-    pos_orient_inds,
-    style=None,
-    **kwargs,
-) -> dict:
-    """
-    Create the plotly mesh3d parameters for a triangle orientation cone or arrow3d in a dictionary
-    based on the provided arguments.
-    """
-    # pylint: disable=protected-access
-    style = obj.style if style is None else style
-    orient = style.orientation
-    size = orient.size
-    symbol = orient.symbol
-    offset = orient.offset
-    color = style.color if orient.color is None else orient.color
-    vertices = obj.mesh if hasattr(obj, "mesh") else [obj.vertices]
-    traces = []
-    for vert in vertices:
-        vec = np.cross(vert[1] - vert[0], vert[2] - vert[1])
-        nvec = vec / np.linalg.norm(vec)
-        # arrow length proportional to square root of triangle
-        length = np.sqrt(triangles_area(np.expand_dims(vert, axis=0))[0]) * 0.2
-        zaxis = np.array([0, 0, 1])
-        cross = np.cross(nvec, zaxis)
-        n = np.linalg.norm(cross)
-        if n == 0:
-            n = 1
-            cross = np.array([-np.sign(nvec[-1]), 0, 0])
-        dot = np.dot(nvec, zaxis)
-        t = np.arccos(dot)
-        vec = -t * cross / n
-        orient = RotScipy.from_rotvec(vec)
-        make_fn = make_BasePyramid if symbol == "cone" else make_BaseArrow
-        vmean = np.mean(vert, axis=0)
-        vmean -= (1 - offset) * length * nvec * size
-        for ind in pos_orient_inds:
-            tr = make_fn(
-                "plotly-dict",
-                base=10,
-                diameter=0.5 * size * length,
-                height=size * length,
-                pivot="tail",
-                color=color,
-                position=obj._orientation[ind].apply(vmean) + obj._position[ind],
-                orientation=obj._orientation[ind] * orient,
-                **kwargs,
-            )
-            traces.append(tr)
-    trace = merge_mesh3d(*traces)
-    return trace
-
-
-def make_Cuboid(
-    obj,
-    style=None,
-    **kwargs,
-) -> dict:
-    """
-    Create the plotly mesh3d parameters for a Cuboid Magnet in a dictionary based on the
-    provided arguments.
-    """
-    style = obj.style if style is None else style
-    dimension = obj.dimension
-    d = [unit_prefix(d / 1000) for d in dimension]
-    trace = make_BaseCuboid("plotly-dict", dimension=dimension, color=style.color)
-    default_suffix = f" ({d[0]}m|{d[1]}m|{d[2]}m)"
-    update_trace_name(trace, "Cuboid", default_suffix, style)
-    return {**trace, **kwargs}
-
-
-def make_Cylinder(
-    obj,
-    style=None,
-    base=50,
-    **kwargs,
-) -> dict:
-    """
-    Create the plotly mesh3d parameters for a Cylinder Magnet in a dictionary based on the
-    provided arguments.
-    """
-    style = obj.style if style is None else style
-    diameter, height = obj.dimension
-    d = [unit_prefix(d / 1000) for d in (diameter, height)]
-    trace = make_BasePrism(
-        "plotly-dict", base=base, diameter=diameter, height=height, color=style.color
-    )
-    default_suffix = f" (D={d[0]}m, H={d[1]}m)"
-    update_trace_name(trace, "Cylinder", default_suffix, style)
-    return {**trace, **kwargs}
-
-
-def make_CylinderSegment(
-    obj,
-    style=None,
-    vertices=25,
-    **kwargs,
-):
-    """
-    Create the plotly mesh3d parameters for a Cylinder Segment Magnet in a dictionary based on the
-    provided arguments.
-    """
-    style = obj.style if style is None else style
-    dimension = obj.dimension
-    d = [unit_prefix(d / (1000 if i < 3 else 1)) for i, d in enumerate(dimension)]
-    trace = make_BaseCylinderSegment(
-        "plotly-dict", dimension=dimension, vert=vertices, color=style.color
-    )
-    default_suffix = f" (r={d[0]}m|{d[1]}m, h={d[2]}m, φ={d[3]}°|{d[4]}°)"
-    update_trace_name(trace, "CylinderSegment", default_suffix, style)
-    return {**trace, **kwargs}
-
-
-def make_Sphere(
-    obj,
-    style=None,
-    vertices=15,
-    **kwargs,
-) -> dict:
-    """
-    Create the plotly mesh3d parameters for a Sphere Magnet in a dictionary based on the
-    provided arguments.
-    """
-    style = obj.style if style is None else style
-    diameter = obj.diameter
-    vertices = min(max(vertices, 3), 20)
-    trace = make_BaseEllipsoid(
-        "plotly-dict", vert=vertices, dimension=[diameter] * 3, color=style.color
-    )
-    default_suffix = f" (D={unit_prefix(diameter / 1000)}m)"
-    update_trace_name(trace, "Sphere", default_suffix, style)
-    return {**trace, **kwargs}
-
-
-def make_Tetrahedron(
-    obj,
-    style=None,
-    **kwargs,
-) -> dict:
-    """
-    Create the plotly mesh3d parameters for a Tetrahedron Magnet in a dictionary based on the
-    provided arguments.
-    """
-    style = obj.style if style is None else style
-    trace = make_BaseTetrahedron(
-        "plotly-dict", vertices=obj.vertices, color=style.color
-    )
-    update_trace_name(trace, "Tetrahedron", "", style)
-    return {**trace, **kwargs}
-
-
-def make_Triangle(
-    obj,
-    style=None,
-    **kwargs,
-) -> dict:
-    """
-    Creates the plotly mesh3d parameters for a Trianglular facet in a dictionary based on the
-    provided arguments.
-    """
-    style = obj.style if style is None else style
-    vert = obj.vertices
-    vec = np.cross(vert[1] - vert[0], vert[2] - vert[1])
-    faces = np.array([[0, 1, 2]])
-    # if magnetization is normal to the triangle, add a second triangle slightly above to enable
-    # proper color gradient visualization. Otherwise only the middle color is shown.
-    if np.all(np.cross(obj.magnetization, vec) == 0):
-        epsilon = 1e-3 * vec
-        vert = np.concatenate([vert - epsilon, vert + epsilon])
-        side_faces = [
-            [0, 1, 3],
-            [1, 2, 4],
-            [2, 0, 5],
-            [1, 4, 3],
-            [2, 5, 4],
-            [0, 3, 5],
-        ]
-        faces = np.concatenate([faces, [[3, 4, 5]], side_faces])
-
-    style = obj.style if style is None else style
-    trace = make_BaseTriangularMesh(
-        "plotly-dict", vertices=vert, faces=faces, color=style.color
-    )
-    update_trace_name(trace, obj.__class__.__name__, "", style)
-    return {**trace, **kwargs}
-
-
-def make_TriangularMesh(
-    obj,
-    color=None,
-    style=None,
-    **kwargs,
-) -> dict:
-    """
-    Creates the plotly mesh3d parameters for a Trianglular facet mesh in a dictionary based on the
-    provided arguments.
-    """
-    style = obj.style if style is None else style
-    trace = make_BaseTriangularMesh(
-        "plotly-dict", vertices=obj.vertices, faces=obj.faces, color=color
-    )
-    ntri = len(obj.faces)
-    default_suffix = f" ({ntri} face{'s'[:ntri^1]})"
-    update_trace_name(trace, obj.__class__.__name__, default_suffix, style)
-    # make edges sharper in plotly
-    trace.update(flatshading=True, lighting_facenormalsepsilon=0, lighting_ambient=0.7)
-    return {**trace, **kwargs}
-
-
-def make_Pixels(positions, size=1) -> dict:
-    """
-    Create the plotly mesh3d parameters for Sensor pixels based on pixel positions and chosen size
-    For now, only "cube" shape is provided.
-    """
-    pixels = [
-        make_BaseCuboid("plotly-dict", position=p, dimension=[size] * 3)
-        for p in positions
-    ]
-    return merge_mesh3d(*pixels)
-
-
-def make_Sensor(
-    obj,
-    style=None,
-    autosize=None,
-    **kwargs,
-):
-    """
-    Create the plotly mesh3d parameters for a Sensor object in a dictionary based on the
-    provided arguments.
-
-    size_pixels: float, default=1
-        A positive number. Adjusts automatic display size of sensor pixels. When set to 0,
-        pixels will be hidden, when greater than 0, pixels will occupy half the ratio of the minimum
-        distance between any pixel of the same sensor, equal to `size_pixel`.
-    """
-    style = obj.style if style is None else style
-    dimension = getattr(obj, "dimension", style.size)
-    pixel = obj.pixel
-    pixel = np.unique(np.array(pixel).reshape((-1, 3)), axis=0)
-    style_arrows = style.arrows.as_dict(flatten=True, separator="_")
-    sensor = get_sensor_mesh(**style_arrows, center_color=style.color)
-    vertices = np.array([sensor[k] for k in "xyz"]).T
-    if style.color is not None:
-        sensor["facecolor"][sensor["facecolor"] == "rgb(238,238,238)"] = style.color
-    dim = np.array(
-        [dimension] * 3 if isinstance(dimension, (float, int)) else dimension[:3],
-        dtype=float,
-    )
-    no_pix = pixel.shape[0] == 1 and (pixel == 0).all()
-    one_pix = pixel.shape[0] == 1 and not (pixel == 0).all()
-    if autosize is not None:
-        dim *= autosize
-    if no_pix:
-        dim_ext = dim
-    else:
-        if one_pix:
-            pixel = np.concatenate([[[0, 0, 0]], pixel])
-        hull_dim = pixel.max(axis=0) - pixel.min(axis=0)
-        dim_ext = max(np.mean(dim), np.min(hull_dim))
-    cube_mask = (vertices < 1).all(axis=1)
-    vertices[cube_mask] = 0 * vertices[cube_mask]
-    vertices[~cube_mask] = dim_ext * vertices[~cube_mask]
-    vertices /= 2  # sensor_mesh vertices are of length 2
-    x, y, z = vertices.T
-    sensor.update(x=x, y=y, z=z)
-    meshes_to_merge = [sensor]
-    if not no_pix:
-        pixel_color = style.pixel.color
-        pixel_size = style.pixel.size
-        combs = np.array(list(combinations(pixel, 2)))
-        vecs = np.diff(combs, axis=1)
-        dists = np.linalg.norm(vecs, axis=2)
-        min_dist = np.min(dists)
-        pixel_dim = dim_ext / 5 if min_dist == 0 else min_dist / 2
-        if pixel_size > 0:
-            pixel_dim *= pixel_size
-            poss = pixel[1:] if one_pix else pixel
-            pixels_mesh = make_Pixels(positions=poss, size=pixel_dim)
-            pixels_mesh["facecolor"] = np.repeat(pixel_color, len(pixels_mesh["i"]))
-            meshes_to_merge.append(pixels_mesh)
-        hull_pos = 0.5 * (pixel.max(axis=0) + pixel.min(axis=0))
-        hull_dim[hull_dim == 0] = pixel_dim / 2
-        hull_mesh = make_BaseCuboid(
-            "plotly-dict", position=hull_pos, dimension=hull_dim
-        )
-        hull_mesh["facecolor"] = np.repeat(style.color, len(hull_mesh["i"]))
-        meshes_to_merge.append(hull_mesh)
-    trace = merge_mesh3d(*meshes_to_merge)
-    default_suffix = (
-        f" ({'x'.join(str(p) for p in obj.pixel.shape[:-1])} pixels)"
-        if obj.pixel.ndim != 1
-        else f" ({pixel[1:].shape[0]} pixel)"
-        if one_pix
-        else ""
-    )
-    update_trace_name(trace, "Sensor", default_suffix, style)
-    return {**trace, **kwargs}
 
 
 def update_magnet_mesh(
@@ -667,20 +108,7 @@ def update_magnet_mesh(
     return mesh_dict
 
 
-def update_trace_name(trace, default_name, default_suffix, style):
-    """provides legend entry based on name and suffix"""
-    name = default_name if style.label is None else style.label
-    if style.description.show and style.description.text is None:
-        name_suffix = default_suffix
-    elif not style.description.show:
-        name_suffix = ""
-    else:
-        name_suffix = f" ({style.description.text})"
-    trace.update(name=f"{name}{name_suffix}")
-    return trace
-
-
-def make_mag_arrows(obj, pos_orient_inds, style):
+def make_mag_arrows(obj):
     """draw direction of magnetization of faced magnets
 
     Parameters
@@ -692,6 +120,7 @@ def make_mag_arrows(obj, pos_orient_inds, style):
     # pylint: disable=protected-access
 
     # vector length, color and magnetization
+    style = obj.style
     if hasattr(obj, "diameter"):
         length = obj.diameter  # Sphere
     elif isinstance(obj, magpy.misc.Triangle):
@@ -705,18 +134,9 @@ def make_mag_arrows(obj, pos_orient_inds, style):
     length *= 1.8 * style.magnetization.size
     mag = obj.magnetization
     # collect all draw positions and directions
-    points = []
-    for ind in pos_orient_inds:
-        pos = getattr(obj, "_barycenter", obj._position)[ind]
-        direc = mag / (np.linalg.norm(mag) + 1e-6) * length
-        vec = obj._orientation[ind].apply(direc)
-        pts = draw_arrowed_line(vec, pos, sign=1, arrow_pos=1, pivot="tail")
-        points.append(pts)
-    # insert empty point to avoid connecting line between arrows
-    points = np.array(points)
-    points = np.insert(points, points.shape[-1], np.nan, axis=2)
-    # remove last nan after insert with [:-1]
-    x, y, z = np.concatenate(points.swapaxes(1, 2))[:-1].T
+    pos = getattr(obj, "_barycenter", obj._position)[0] - obj._position[0]
+    direc = mag / (np.linalg.norm(mag) + 1e-6) * length
+    x, y, z = draw_arrowed_line(direc, pos, sign=1, arrow_pos=1, pivot="tail").T
     trace = {
         "type": "scatter3d",
         "mode": "lines",
@@ -730,8 +150,9 @@ def make_mag_arrows(obj, pos_orient_inds, style):
     return trace
 
 
-def make_path(input_obj, style):
+def make_path(input_obj):
     """draw obj path based on path style properties"""
+    style = input_obj.style
     x, y, z = np.array(input_obj.position).T
     txt_kwargs = (
         {"mode": "markers+text+lines", "text": list(range(len(x)))}
@@ -885,7 +306,7 @@ def get_generic_traces_2D(
     def get_label_and_color(obj):
         props = flat_objs_props.get(obj, {})
         style = props.get("style", None)
-        style = obj.style if style is None else style
+        style = obj.style
         label = getattr(style, "label", None)
         label = repr(obj) if not label else label
         color = getattr(style, "color", None)
@@ -958,7 +379,6 @@ def get_generic_traces(
     extra_backend=False,
     row=1,
     col=1,
-    style=None,
     **kwargs,
 ) -> list:
     """
@@ -981,15 +401,9 @@ def get_generic_traces(
     # pylint: disable=protected-access
     # pylint: disable=import-outside-toplevel
 
-    from magpylib._src.obj_classes.class_misc_Triangle import Triangle
-    from magpylib._src.obj_classes.class_magnet_TriangularMesh import TriangularMesh
-
-    # parse kwargs into style and non style args
-
+    style = input_obj.style
     is_mag_arrows = False
-    is_mag = hasattr(input_obj, "magnetization") and hasattr(
-        input_obj.style, "magnetization"
-    )
+    is_mag = hasattr(input_obj, "magnetization") and hasattr(style, "magnetization")
     if is_mag and style.magnetization.show:
         mag = style.magnetization
         if mag.mode == "auto":
@@ -997,183 +411,116 @@ def get_generic_traces(
         is_mag_arrows = "arrow" in mag.mode
         mag.show = "color" in mag.mode
 
-    make_func = input_obj._draw_func
-    make_func_kwargs = {"style": style, **kwargs}
+    make_func = getattr(input_obj, "get_trace", None)
+    make_func_kwargs = {"legendgroup": legendgroup, **kwargs}
     if getattr(input_obj, "_autosize", False):
         make_func_kwargs["autosize"] = autosize
 
-    all_generic_traces = []
-    path_traces = []
-    path_traces_extra_generic = []
-    path_traces_extra_specific_backend = []
     has_path = hasattr(input_obj, "position") and hasattr(input_obj, "orientation")
-    if not has_path:
+    path_traces_extra_non_generic_backend = []
+    if not has_path and make_func is not None:
         tr = make_func(**make_func_kwargs)
         tr["row"] = row
         tr["col"] = col
         out = {"generic": [tr]}
         if extra_backend:
-            out.update({extra_backend: path_traces_extra_specific_backend})
+            out.update({extra_backend: path_traces_extra_non_generic_backend})
         return out
 
-    extra_model3d_traces = style.model3d.data if style.model3d.data is not None else []
     orientations, positions, pos_orient_inds = get_rot_pos_from_path(
         input_obj, style.path.frames
     )
-    obj_is_disconnected = False
-    if isinstance(input_obj, TriangularMesh):
-        for mode in ("open", "disconnected", "selfintersecting"):
-            show_mesh = getattr(style.mesh, mode).show
-            if mode == "open" and show_mesh:
-                if input_obj.status_open is None:
-                    warnings.warn(
-                        f"Unchecked open mesh status in {input_obj!r} detected, before attempting "
-                        "to show potential open edges, which may take a while to compute "
-                        "when the mesh has many faces, now applying operation..."
-                    )
-                    input_obj.check_open()
-            elif mode == "disconnected" and show_mesh:
-                if input_obj.status_disconnected is None:
-                    warnings.warn(
-                        f"Unchecked disconnected mesh status in {input_obj!r} detected, before "
-                        "attempting to show possible disconnected parts, which may take a while "
-                        "to compute when the mesh has many faces, now applying operation..."
-                    )
-                obj_is_disconnected = input_obj.check_disconnected()
-            elif mode == "selfintersecting":
-                if input_obj._status_selfintersecting is None:
-                    warnings.warn(
-                        f"Unchecked selfintersecting mesh status in {input_obj!r} detected, before "
-                        "attempting to show possible disconnected parts, which may take a while "
-                        "to compute when the mesh has many faces, now applying operation..."
-                    )
-                    input_obj.check_selfintersecting()
-    disconnected_traces = []
-    for orient, pos in zip(orientations, positions):
+    traces_generic = []
+    if pos_orient_inds.size != 0:
         if style.model3d.showdefault and make_func is not None:
-            if obj_is_disconnected:
-                tria_orig = input_obj._faces
-                mag_show = style.magnetization.show
-                for tri, dis_color in zip(
-                    input_obj.get_faces_subsets(),
-                    cycle(style.mesh.disconnected.colorsequence),
-                ):
-                    # temporary mutate faces from subset
-                    input_obj._faces = tri
-                    style.magnetization.show = False
-                    dis_tr = make_func(
-                        **{**make_func_kwargs, "color": dis_color},
+            p_trs = make_func(**make_func_kwargs)
+            p_trs = [p_trs] if isinstance(p_trs, dict) else p_trs
+            for p_tr in p_trs:
+                is_mag = p_tr.pop("ismagnet", is_mag)
+                if is_mag and p_tr.get("type", "") == "mesh3d":
+                    p_tr = update_magnet_mesh(
+                        p_tr,
+                        mag_style=style.magnetization,
+                        magnetization=input_obj.magnetization,
+                        color_slicing=not supports_colorgradient,
                     )
-                    dis_tr = place_and_orient_model3d(
-                        dis_tr, orientation=orient, position=pos
-                    )
-                    disconnected_traces.append(dis_tr)
-                input_obj._faces = tria_orig
-                style.magnetization.show = mag_show
-            else:  # if disconnnected, no mag slicing needed
-                p_trs = make_func(**make_func_kwargs)
-                p_trs = [p_trs] if isinstance(p_trs, dict) else p_trs
-                for p_tr in p_trs:
-                    if is_mag:
-                        p_tr = update_magnet_mesh(
-                            p_tr,
-                            mag_style=style.magnetization,
-                            magnetization=input_obj.magnetization,
-                            color_slicing=not supports_colorgradient,
-                        )
-                    p_tr = place_and_orient_model3d(
-                        p_tr, orientation=orient, position=pos
-                    )
-                    path_traces.append(p_tr)
+
+                traces_generic.append(p_tr)
+
+        extra_model3d_traces = (
+            style.model3d.data if style.model3d.data is not None else []
+        )
         for extr in extra_model3d_traces:
-            if extr.show:
+            if not extr.show:
+                continue
+            extr.update(extr.updatefunc())  # update before checking backend
+            if extr.backend == "generic":
                 extr.update(extr.updatefunc())
-                if extr.backend == "generic":
-                    trace3d = {"opacity": style.opacity}
-                    ttype = extr.constructor.lower()
-                    obj_extr_trace = (
-                        extr.kwargs() if callable(extr.kwargs) else extr.kwargs
-                    )
-                    obj_extr_trace = {"type": ttype, **obj_extr_trace}
-                    if ttype == "scatter3d":
-                        for k in ("marker", "line"):
-                            trace3d[f"{k}_color"] = trace3d.get(
-                                f"{k}_color", style.color
-                            )
-                    elif ttype == "mesh3d":
-                        trace3d["showscale"] = trace3d.get("showscale", False)
-                        if "facecolor" in obj_extr_trace:
-                            ttype = "mesh3d_facecolor"
-                        trace3d["color"] = trace3d.get("color", style.color)
-                    else:  # pragma: no cover
-                        raise ValueError(
-                            f"{ttype} is not supported, only 'scatter3d' and 'mesh3d' are"
+                tr_generic = {"opacity": style.opacity}
+                ttype = extr.constructor.lower()
+                obj_extr_trace = extr.kwargs() if callable(extr.kwargs) else extr.kwargs
+                obj_extr_trace = {"type": ttype, **obj_extr_trace}
+                if ttype == "scatter3d":
+                    for k in ("marker", "line"):
+                        tr_generic[f"{k}_color"] = tr_generic.get(
+                            f"{k}_color", style.color
                         )
-                    trace3d.update(
-                        linearize_dict(
-                            place_and_orient_model3d(
-                                model_kwargs=obj_extr_trace,
-                                orientation=orient,
-                                position=pos,
-                                scale=extr.scale,
-                            ),
-                            separator="_",
-                        )
+                elif ttype == "mesh3d":
+                    tr_generic["showscale"] = tr_generic.get("showscale", False)
+                    if "facecolor" in obj_extr_trace:
+                        ttype = "mesh3d_facecolor"
+                    tr_generic["color"] = tr_generic.get("color", style.color)
+                else:  # pragma: no cover
+                    raise ValueError(
+                        f"{ttype} is not supported, only 'scatter3d' and 'mesh3d' are"
                     )
-                    path_traces_extra_generic.append(trace3d)
-                elif extr.backend == extra_backend:
-                    trace3d = {
-                        "model3d": extr,
-                        "position": pos,
-                        "orientation": orient,
-                        "kwargs": {
-                            "opacity": style.opacity,
-                            "color": style.color,
-                            "legendgroup": legendgroup,
-                            "name": legendtext,
-                            "row": row,
-                            "col": col,
-                        },
-                    }
-                    path_traces_extra_specific_backend.append(trace3d)
-
-    all_generic_traces.extend(group_traces(*path_traces, *path_traces_extra_generic))
-
-    if disconnected_traces:
-        nsubsets = len(input_obj.get_faces_subsets())
-        for ind in range(nsubsets):
-            trace = merge_traces(*disconnected_traces[ind::nsubsets])[0]
-            trace["legendgroup"] = f"{legendgroup} - part_{ind+1:02d}"
-            lg = trace.get("name", "") if legendtext is None else legendtext
-            trace["name"] = f"{lg} - part_{ind+1:02d}"
-            all_generic_traces.append(trace)
-
-    if np.array(input_obj.position).ndim > 1 and style.path.show:
-        scatter_path = make_path(input_obj, style)
-        all_generic_traces.append(scatter_path)
+                tr_generic.update(linearize_dict(obj_extr_trace, separator="_"))
+                traces_generic.append(tr_generic)
 
     if is_mag_arrows:
-        all_generic_traces.append(make_mag_arrows(input_obj, pos_orient_inds, style))
-    if isinstance(input_obj, (Triangle, TriangularMesh)) and style.orientation.show:
-        all_generic_traces.append(
-            make_triangle_orientations(
-                input_obj,
-                pos_orient_inds,
-                style=style,
-                legendgroup=legendgroup,
-                **kwargs,
-            )
-        )
-    if isinstance(input_obj, TriangularMesh):
-        for mode in ("grid", "open", "disconnected", "selfintersecting"):
-            if getattr(style.mesh, mode).show:
-                trace = make_mesh_lines(
-                    input_obj, pos_orient_inds, mode, style, **kwargs
-                )
-                if trace:
-                    all_generic_traces.append(trace)
+        mag_arrow_tr = make_mag_arrows(input_obj)
+        traces_generic.append(mag_arrow_tr)
 
-    for tr in all_generic_traces:
+    path_traces_generic = []
+    for tr in traces_generic:
+        temp_rot_traces = []
+        name_suff = tr.pop("name_suffix", None)
+        name = tr.get("name", "") if legendtext is None else legendtext
+        for orient, pos in zip(orientations, positions):
+            tr1 = place_and_orient_model3d(tr, orientation=orient, position=pos)
+            if name_suff is not None:
+                tr1["name"] = f"{name}{name_suff}"
+            temp_rot_traces.append(tr1)
+        path_traces_generic.extend(group_traces(*temp_rot_traces))
+
+    for extr in extra_model3d_traces:
+        if not extr.show:
+            continue
+        extr.update(extr.updatefunc())  # update before checking backend
+        if extr.backend == extra_backend:
+            for orient, pos in zip(orientations, positions):
+                tr_generic = {
+                    "model3d": extr,
+                    "position": pos,
+                    "orientation": orient,
+                    "kwargs": {
+                        "opacity": style.opacity,
+                        "color": style.color,
+                        "legendgroup": legendgroup,
+                        "name": legendtext,
+                        "row": row,
+                        "col": col,
+                    },
+                }
+                path_traces_extra_non_generic_backend.append(tr_generic)
+
+    if np.array(input_obj.position).ndim > 1 and style.path.show:
+        scatter_path = make_path(input_obj)
+        path_traces_generic.append(scatter_path)
+
+    path_traces_generic.extend(group_traces(*path_traces_generic))
+
+    for tr in path_traces_generic:
         tr.update(row=row, col=col)
         if tr.get("opacity", None) is None:
             tr["opacity"] = style.opacity
@@ -1187,9 +534,9 @@ def get_generic_traces(
         if tr.get("facecolor", None) is not None:
             # this allows merging of 3d meshes, ignoring different colors
             tr["color"] = None
-    out = {"generic": all_generic_traces}
+    out = {"generic": path_traces_generic}
     if extra_backend:
-        out.update({extra_backend: path_traces_extra_specific_backend})
+        out.update({extra_backend: path_traces_extra_non_generic_backend})
     return out
 
 
@@ -1307,13 +654,7 @@ def extract_animation_properties(
     return path_indices, exp, frame_duration
 
 
-def draw_frame(
-    objs,
-    colorsequence=None,
-    zoom=0.0,
-    autosize=None,
-    **kwargs,
-) -> Tuple:
+def draw_frame(objs, colorsequence=None, zoom=0.0, autosize=None, **kwargs) -> Tuple:
     """
     Creates traces from input `objs` and provided parameters, updates the size of objects like
     Sensors and Dipoles in `kwargs` depending on the canvas size.
@@ -1378,9 +719,19 @@ def get_row_col_traces(flat_objs_props, extra_backend=False, autosize=None, **kw
             for rco in rco_obj:
                 params["row"], params["col"], output_typ = rco
                 if output_typ == "model3d":
-                    out_traces = get_generic_traces(
-                        obj, extra_backend=extra_backend, autosize=autosize, **params
-                    )
+                    orig_style = None
+                    try:
+                        # temporary replace style attribute
+                        orig_style = obj._style
+                        obj._style = params.pop("style", None)
+                        out_traces = get_generic_traces(
+                            obj,
+                            extra_backend=extra_backend,
+                            autosize=autosize,
+                            **params,
+                        )
+                    finally:
+                        obj._style = orig_style
                     if extra_backend:
                         extra_backend_traces.extend(out_traces.get(extra_backend, []))
                     traces_dict[obj].extend(out_traces["generic"])
