+++ magpylib/_src/obj_classes/class_BaseDisplayRepr.py
@@ -4,7 +4,7 @@
 import numpy as np
 
 from magpylib._src.display.display import show
-from magpylib._src.display.traces_generic import make_DefaultTrace
+from magpylib._src.display.traces_core import make_DefaultTrace
 
 UNITS = {
     "parent": None,
@@ -21,7 +21,7 @@ class BaseDisplayRepr:
     """Provides the show and repr methods for all objects"""
 
     show = show
-    _draw_func = make_DefaultTrace
+    get_trace = make_DefaultTrace
 
     def _property_names_generator(self):
         """returns a generator with class properties only"""
