+++ magpylib/_src/obj_classes/class_current_Line.py
@@ -1,5 +1,5 @@
 """Line current class code"""
-from magpylib._src.display.traces_generic import make_Line
+from magpylib._src.display.traces_core import make_Line
 from magpylib._src.fields.field_BH_line import current_vertices_field
 from magpylib._src.input_checks import check_format_input_vertices
 from magpylib._src.obj_classes.class_BaseExcitations import BaseCurrent
@@ -89,7 +89,7 @@ class Line(BaseCurrent):
         "segment_start": 2,
         "segment_end": 2,
     }
-    _draw_func = make_Line
+    get_trace = make_Line
 
     def __init__(
         self,
