+++ magpylib/_src/obj_classes/class_magnet_CylinderSegment.py
@@ -1,7 +1,7 @@
 """Magnet Cylinder class code"""
 import numpy as np
 
-from magpylib._src.display.traces_generic import make_CylinderSegment
+from magpylib._src.display.traces_core import make_CylinderSegment
 from magpylib._src.fields.field_BH_cylinder_segment import (
     magnet_cylinder_segment_field_internal,
 )
@@ -93,7 +93,7 @@ class CylinderSegment(BaseMagnet):
 
     _field_func = staticmethod(magnet_cylinder_segment_field_internal)
     _field_func_kwargs_ndim = {"magnetization": 2, "dimension": 2}
-    _draw_func = make_CylinderSegment
+    get_trace = make_CylinderSegment
 
     def __init__(
         self,
