+++ magpylib/_src/obj_classes/class_magnet_Sphere.py
@@ -1,5 +1,5 @@
 """Magnet Sphere class code"""
-from magpylib._src.display.traces_generic import make_Sphere
+from magpylib._src.display.traces_core import make_Sphere
 from magpylib._src.fields.field_BH_sphere import magnet_sphere_field
 from magpylib._src.input_checks import check_format_input_scalar
 from magpylib._src.obj_classes.class_BaseExcitations import BaseMagnet
@@ -80,7 +80,7 @@ class Sphere(BaseMagnet):
 
     _field_func = staticmethod(magnet_sphere_field)
     _field_func_kwargs_ndim = {"magnetization": 2, "diameter": 1}
-    _draw_func = make_Sphere
+    get_trace = make_Sphere
 
     def __init__(
         self,
