+++ magpylib/_src/obj_classes/class_misc_Dipole.py
@@ -1,5 +1,5 @@
 """Dipole class code"""
-from magpylib._src.display.traces_generic import make_Dipole
+from magpylib._src.display.traces_core import make_Dipole
 from magpylib._src.fields.field_BH_dipole import dipole_field
 from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.obj_classes.class_BaseExcitations import BaseSource
@@ -79,7 +79,7 @@ class Dipole(BaseSource):
     _field_func = staticmethod(dipole_field)
     _field_func_kwargs_ndim = {"moment": 2}
     _style_class = DipoleStyle
-    _draw_func = make_Dipole
+    get_trace = make_Dipole
     _autosize = True
 
     def __init__(
