+++ magpylib/_src/style.py
@@ -15,7 +15,7 @@ from magpylib._src.defaults.defaults_utility import validate_style_keys
 
 
 def get_families(obj):
-    "get obj families"
+    """get obj families"""
     # pylint: disable=import-outside-toplevel
     # pylint: disable=possibly-unused-variable
     # pylint: disable=redefined-outer-name
@@ -209,9 +209,12 @@ class BaseStyle(MagicProperties):
 
     @description.setter
     def description(self, val):
-        self._description = validate_property_class(
-            val, "description", Description, self
-        )
+        if isinstance(val, str):
+            self._description = Description(text=val)
+        else:
+            self._description = validate_property_class(
+                val, "description", Description, self
+            )
 
     @property
     def color(self):
